requirements.txt
fastapi
jinja2
python-dotenv
uvicorn